proyecto ITRN

+++ Nuclear-Teorica/ITRN/c++/OpticModel/Numerov/Numerov.h
@@ -4,11 +4,11 @@
 #include <complex>
 
 class Potential;
-using Complex = std::complex<double>;
+using Complex = std::complex<long double>;
 
 // Estructura para resultados complejos
 struct ComplexNumerovResult {
-    std::vector<double> r;
+    std::vector<long double> r;
     std::vector<Complex> u;
     std::vector<Complex> du;
     
@@ -20,9 +20,9 @@ struct ComplexNumerovResult {
 };
 
 // Función principal para potenciales complejos
-ComplexNumerovResult numerov_complex(double l, double E, double a, const Potential& pot, double mua = 0.0);
+ComplexNumerovResult numerov_complex(long double l, long double E, long double a, const Potential& pot, long double mua = 0.0);
 
 // Función de compatibilidad (solo parte real)
-std::pair<std::vector<double>, std::vector<double>> numerov(double l, double E, double a, const Potential& pot, double mua = 0.0);
+std::pair<std::vector<long double>, std::vector<long double>> numerov(long double l, long double E, long double a, const Potential& pot, long double mua = 0.0);
 
 #endif
